refactor(analytics): move app/analytics/ under app/core/analytics/
