+++ app/core/analytics/schemas/base.py
