+++ app/core/analytics/schemas/v1/user_logged_in.py
@@ -4,7 +4,7 @@ Emitted fire-and-forget by the login endpoint on a successful password login —
 the first real producer wired to the emission gateway.
 """
 
-from app.analytics.schemas.base import AnalyticsEventSchema
+from app.core.analytics.schemas.base import AnalyticsEventSchema
 
 
 class UserLoggedIn(AnalyticsEventSchema):
