+++ app/testing.py
@@ -32,7 +32,7 @@ from httpx import ASGITransport, AsyncClient
 from sqlmodel import SQLModel
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-import app.analytics.db as analytics_db
+import app.core.analytics.db as analytics_db
 import app.core.db as core_db
 import app.lib.db as db
 from app.api.v1.auth import get_current_user
@@ -80,7 +80,7 @@ async def _db_schema(monkeypatch: pytest.MonkeyPatch) -> AsyncGenerator[None]:
     async def _open_analytics_session(expire_on_commit: bool = False) -> AsyncGenerator[AsyncSession]:
         nonlocal analytics_schema_created
         if not analytics_schema_created:
-            from app.analytics.models import analytics_metadata
+            from app.core.analytics.models import analytics_metadata
 
             async with analytics_db.get_analytics_engine().begin() as conn:
                 await conn.run_sync(lambda c: analytics_metadata.create_all(c, checkfirst=False))
