+++ tests/analytics/test_events_endpoint.py
@@ -4,7 +4,7 @@ from httpx import AsyncClient
 from sqlalchemy import func, select
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from app.analytics.models import raw_events
+from app.core.analytics.models import raw_events
 from app.models.user import User
 
 URL = "/api/v1/events/"
